fix: rename transactions page files
